# openapi/social-backend.openapi.yaml
openapi: 3.0.3
info:
  title: Social Backend API
  version: 1.0.0
  license:
    name: UNLICENSED
  description: |
    Public API contract exposed through the KrakenD gateway.
    Frontend clients should use these gateway paths, not internal service paths.
servers:
  - url: http://localhost:8080
    description: Local KrakenD gateway
...
security:
  - bearerAuth: []
paths:
...
  /api/messages/chats:
    get:
      tags: [Messages]
      summary: List chats for current user
      operationId: listChats
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Chat list
    post:
      tags: [Messages]
      summary: Create or return existing 1-1 chat
      operationId: createChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [participantIds]
              properties:
                participantIds:
                  type: array
                  items: { type: integer }
                isGroup: { type: boolean }
                chatName: { type: string }
      responses:
        '200':
          description: Flat ChatData
  /api/messages/chats/{chatId}:
    delete:
      tags: [Messages]
      summary: Delete chat
      operationId: deleteChat
      parameters:
        - name: chatId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Deleted
  /api/messages/chats/{chatId}/upload-url:
    get:
      tags: [Messages]
      summary: Get presigned URL for message attachment upload
      operationId: getMessageAttachmentUploadUrl
      parameters:
        - name: chatId
          in: path
          required: true
          schema: { type: integer }
        - name: contentType
          in: query
          schema: { type: string }
        - name: fileName
          in: query
          schema: { type: string }
        - name: sizeBytes
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Presigned upload URL payload
  /api/messages/chats/{chatId}/attachments:
    get:
      tags: [Messages]
      summary: List chat attachments (photos and files)
      operationId: listChatAttachments
      parameters:
        - name: chatId
          in: path
          required: true
          schema: { type: integer }
        - name: kind
          in: query
          schema: { type: string, enum: [image, file] }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Attachments list
  /api/messages/send:
    post:
      tags: [Messages]
      summary: Send a message (text and/or attachments)
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [chatId]
              properties:
                chatId: { type: integer }
                content: { type: string, maxLength: 4000 }
                attachments:
                  type: array
                  maxItems: 5
                  items:
                    type: object
                    required: [url, fileName, mimeType, sizeBytes, objectKey]
                    properties:
                      url: { type: string, format: uri }
                      fileName: { type: string }
                      mimeType: { type: string }
                      sizeBytes: { type: integer }
                      objectKey: { type: string }
      responses:
        '200':
          description: Flat MessageData with attachments
  /api/messages/forward:
    post:
      tags: [Messages]
      summary: Forward messages to one or more chats
      operationId: forwardMessages
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messageIds, targetChatIds]
              properties:
                messageIds:
                  type: array
                  items: { type: integer }
                  minItems: 1
                  maxItems: 50
                targetChatIds:
                  type: array
                  items: { type: integer }
                  minItems: 1
                  maxItems: 20
      responses:
        '200':
          description: Forwarded MessageData list
  /api/messages/bulk:
    delete:
      tags: [Messages]
      summary: Bulk delete own messages
      operationId: deleteMessagesBulk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messageIds]
              properties:
                messageIds:
                  type: array
                  items: { type: integer }
                  minItems: 1
                  maxItems: 100
      responses:
        '200':
          description: Deleted ids
  /api/messages/{chatId}:
    get:
      tags: [Messages]
      summary: List messages (marks chat as read)
      operationId: listMessages
      parameters:
        - name: chatId
          in: path
          required: true
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: before
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Messages list with attachments
    patch:
      tags: [Messages]
      summary: Edit own message content and attachments
      operationId: editMessage
      parameters:
        - name: chatId
          in: path
          required: true
          description: Message id
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content: { type: string, maxLength: 4000 }
                removeAttachmentIds:
                  type: array
                  items: { type: integer }
                  maxItems: 50
                attachments:
                  type: array
                  maxItems: 5
                  items:
                    type: object
                    required: [url, fileName, mimeType, sizeBytes, objectKey]
                    properties:
                      url: { type: string }
                      fileName: { type: string }
                      mimeType: { type: string }
                      sizeBytes: { type: integer }
                      objectKey: { type: string }
      responses:
        '200':
          description: Updated MessageData
    delete:
      tags: [Messages]
      summary: Delete own message
      operationId: deleteMessage
      parameters:
        - name: chatId
          in: path
          required: true
          description: Message id
          schema: { type: integer }
      responses:
        '200':
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...